pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless==4.10.0.82",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
]
